Exhibit 1: Ukraine and Russia War Conflict Pattern
Exhibit 1: Ukraine and Russia War Conflict Pattern

Ukraine War is Planned by the WEF World Economic Forum Members.
Because On January 16 I Emailed the Crimes About the Jaggi Vasudev and PM Care Scandals to the United States White House, President of the United States and Including Vice President Kamala Harris.

After the Few Days Before or on January 20 2022 the Corrupt Leader and Genocider and One of the Members of the World Economic Forum.


Bill Gates Met Imran Khan on February 17, 2022, the Reason was told is Polio Eradication. But the truth is He Convince the Imran Khan to Vist the Putin.

He Told the Media that Someone is Trying to Trannish the India Image at Global Level.

Because They Feared about that Mail I Sent. On February 6, 2022, my novecore account was Hacked intentionally by Some Government I Don't Like Mention Who is that? 

February 7, 2022 - Macron Visited the Putin.

February 18 - 23, They Planned to Start the War, Munich Security Conference 2022.

The Total Scenario and Action of them to start the War and Divert the Peoples Attention and Cover UP the Genocide Vaccine Side Effects.

This War is Planned to Cover UP the Vaccine Side Effects.